Rename calc file to OneRepMax
